refactor: split StructAccess
